FIX PLAYER/SHIP PREFAB
Restructured ship prefab in two GOs, one for physics and scripts, the
other for graphics, i.e. sprite and animator. Also set the animator as
private in PlayerController, as currently it was visible in the
inspector (empty), but then GetComponent is called in the Start method.
Since the animator is now a child object I also switched out the method
for GetComponentInChildren. Lastly, I tested various colliders and the
simplest (i.e. best performing) and most natural feeling (tested
crashing into asteroids) was the polygon collider. PS! Prefab did not
have a collider attached to it before.