FIX PROJECTILE PLAYER COLLISION BUG
Set the projectile layer to player and removed collisions between
objects in the player layer in the physics settings. Was causing the
projectile to register a collision with the ship and caused it to spin.
The bug finally showed itself since a large enough collider was added
to the ship prefab.