FIX ASTEROID PREFAB
Restructured asteroid prefab as other prefabs with physics/logic and
graphics on separate GO-s, the second GO being the child of the first.
Also created a separate sorting layer for destructibles and added it
to the prefabs sprite renderer. While I was at it, I took the liberty of
rearranging the sorting layers to render as intended.