FIX PLAYER HIT LOGIC IMPLEMENTATION
The player is not a destructible object, meaning it's logic differs too
widely from a destructible object and will only differ further as
development progresses. The intent is not the same. Moved logic out
of Player script and into PlayerController, this also avoids similar
script name proliferation. Left the player death logic unimplemented on
purpose, nothing to tie it to yet.