REMOVE ASTEROID MASS
While it's set it does not get any functional use currently.
Destructibles don't collide with each other and the player dies with a
single hit, i.e. before anything is simulated. It's also not used when
applying a movement force to the asteroid and shouldn't be, in order to
not complicate the simulation.