REFACTOR ASTEROID DESPAWN LOGIC
Whilst simpler the gameobject lifetime variable, used together with
Destroy method is unreliable. The lifetime would need to be rather long
in order to ensure that the object is truly out of view, since we'd
need to account for the play area size and the speed of the object.
Introduced a DespawnIfNotVisible method instead that simply checks
whether the object is visible to any camera. If not, despawn.

PS! This includes the scene camera, check documentation:
https://docs.unity3d.com/ScriptReference/Renderer-isVisible.html