REFACTOR SPAWN LOGIC
Refactored the spawn logic in order to isolate the spawn method as
completely as possible from the actual spawned object. Uses two points
on two circles, calculate the angle between the two and rotate the
Instantiated object towards the destination point. This way all movement
logic can be handled separately inside the destructible object since
the trajectory is set.